refactor: improve code and rename variables for clarity

+++ internal/handlers/collections/delete_collection.go
@@ -2,56 +2,38 @@ package collections
 
 import (
 	"github.com/k4sper1love/watchlist-bot/internal/builders/keyboards"
+	"github.com/k4sper1love/watchlist-bot/internal/builders/messages"
 	"github.com/k4sper1love/watchlist-bot/internal/handlers/states"
 	"github.com/k4sper1love/watchlist-bot/internal/models"
 	"github.com/k4sper1love/watchlist-bot/internal/services/watchlist"
 	"github.com/k4sper1love/watchlist-bot/internal/utils"
-	"github.com/k4sper1love/watchlist-bot/pkg/translator"
 )
 
 func HandleDeleteCollectionCommand(app models.App, session *models.Session) {
-	msg := "⚠️" + translator.Translate(session.Lang, "deleteCollectionConfirm", map[string]interface{}{
-		"Collection": session.CollectionDetailState.Collection.Name,
-	}, nil)
-
-	keyboard := keyboards.NewKeyboard().AddSurvey().Build(session.Lang)
-
-	app.SendMessage(msg, keyboard)
-
-	session.SetState(states.ProcessDeleteCollectionAwaitingConfirm)
+	app.SendMessage(messages.DeleteCollection(session), keyboards.Survey(session))
+	session.SetState(states.AwaitDeleteCollectionConfirm)
 }
 
 func HandleDeleteCollectionProcess(app models.App, session *models.Session) {
 	switch session.State {
-	case states.ProcessDeleteCollectionAwaitingConfirm:
+	case states.AwaitDeleteCollectionConfirm:
 		parseDeleteCollectionConfirm(app, session)
+		session.ClearState()
 	}
 }
 
 func parseDeleteCollectionConfirm(app models.App, session *models.Session) {
-	session.ClearState()
-
-	switch utils.IsAgree(app.Upd) {
-	case true:
-		if err := watchlist.DeleteCollection(app, session); err != nil {
-			msg := "🚨 " + translator.Translate(session.Lang, "deleteCollectionFailure", map[string]interface{}{
-				"Collection": session.CollectionDetailState.Collection.Name,
-			}, nil)
-			keyboard := keyboards.NewKeyboard().AddBack(states.CallbackCollectionsManage).Build(session.Lang)
-			app.SendMessage(msg, keyboard)
-			break
-		}
-
-		msg := "🗑️ " + translator.Translate(session.Lang, "deleteCollectionSuccess", map[string]interface{}{
-			"Collection": session.CollectionDetailState.Collection.Name,
-		}, nil)
-
-		app.SendMessage(msg, nil)
-		HandleCollectionsCommand(app, session)
-
-	case false:
-		msg := "🚫 " + translator.Translate(session.Lang, "cancelAction", nil, nil)
-		app.SendMessage(msg, nil)
+	if !utils.IsAgree(app.Update) {
+		app.SendMessage(messages.CancelAction(session), nil)
 		HandleManageCollectionCommand(app, session)
+		return
 	}
+
+	if err := watchlist.DeleteCollection(app, session); err != nil {
+		app.SendMessage(messages.DeleteCollectionFailure(session), keyboards.Back(session, states.CallCollectionsManage))
+		return
+	}
+
+	app.SendMessage(messages.DeleteCollectionSuccess(session), nil)
+	HandleCollectionsCommand(app, session)
 }
